Fix data is ignored because effect rerun
Because of effect will run the cleanup function before
the fetch is completed. the ignore flag for that effect
is already true, that why it got ignored. if i want to use ignore flag
i need to await for fetch to be done first before running the
clean up that set the ignore to true